pyproject.toml
[project]
name = "kafka-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiokafka>=0.13.0",
    "asyncio>=4.0.0",
    "confluent-kafka[avro,schemaregistry]>=2.6.0",
    "fastavro>=1.9.0",
    "kafka-python>=2.3.0",
    "mcp>=1.25.0",
    "python-dotenv>=1.2.1",
    "setuptools>=80.9.0",
    "jsonschema>=4.0.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
agent = [
    "langchain>=1.2.3",
    "langchain-core>=1.2.7",
    "langchain-groq>=1.1.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.7",
    "langgraph>=0.2.0",
    "mcp-server-fetch>=2025.4.7",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "pytest-timeout>=2.0.0",
    "starlette>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["test"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
markers = [
    "integration: tests requiring Docker Kafka and/or Schema Registry",
    "llm: tests requiring real LLM API calls (OPENAI_API_KEY)",
    "schema_registry: tests requiring Schema Registry",
]

[tool.coverage.run]
source = ["src"]
omit = ["test/*", ".venv/*", "src/main.py", "src/client.py", "src/prompt_template.py"]
